Crate pliantdb_core

Core functionality and types for PliantDB.

Re-exports

pub use circulate;

pub use fabruic;

pub use num_traits;

Modules

connection
Types for interacting with a database.

document
Types for interacting with Documents.

limits
Limits used within PliantDB.

networking
Types for networked communications.

pubsub
Types for Publish/Subscribe (PubSub) messaging.

schema
Types for defining database schema.

test_util
Shared schemas and utilities used for unit testing.

transaction
Types for executing transactions.

Macros

define_connection_test_suite
Creates a test suite that tests methods available on Connection

define_pubsub_test_suite
Expands into a suite of pubsub unit tests using the passed type as the test harness.

Enums

Error
an enumeration of errors that this crate can produce