Trait pliantdb_core::backend::Backend

pub trait Backend: Debug + Send + Sync + 'static {
    type CustomApi: CustomApi;
}

Tailors the behavior of a server to your needs.

Associated Types

type CustomApi: CustomApi

The custom API definition. If you do not wish to have an API, () may be provided.

Loading content...