Enum pliantdb_core::Error

pub enum Error {

    SchemaMismatch {
        database_name: String,
        schema: SchemaName,
        stored_schema: SchemaName,
    },
    SchemaAlreadyRegistered(SchemaName),
    SchemaNotRegistered(SchemaName),
    InvalidDatabaseName(String),
    DatabaseNotFound(String),
    DatabaseNameAlreadyTaken(String),
    Database(String),
    Server(String),
    Transport(String),
    Websocket(String),
    Networking(Error),
    Io(String),
    Configuration(String),
    Client(String),
    CollectionNotFound,
    CollectionAlreadyDefined,
    DocumentNotFound(CollectionName, u64),
    DocumentConflict(CollectionName, u64),
    UniqueKeyViolation {
        view: ViewName,
        conflicting_document_id: u64,
        existing_document_id: u64,
    },
    InvalidName(InvalidNameError),
    PermissionDenied(PermissionDenied),
    Password(String),
    UserNotFound,
    InvalidCredentials,

}

an enumeration of errors that this crate can produce

Variants

SchemaMismatch

The database named database_name was created with a different schema (stored_schema) than provided (schema).

Fields of SchemaMismatch

database_name: String

The name of the database being accessed.

schema: SchemaName

The schema provided for the database.

stored_schema: SchemaName

The schema stored for the database.

SchemaAlreadyRegistered(SchemaName)

The SchemaName returned has already been registered.

SchemaNotRegistered(SchemaName)

The SchemaName requested was not registered.

InvalidDatabaseName(String)

An invalid database name was specified. See ServerConnection::create_database() for database name requirements.

DatabaseNotFound(String)

The database name given was not found.

DatabaseNameAlreadyTaken(String)

The database name already exists.

Database(String)

An error from interacting with local storage.

Server(String)

An error from interacting with a server.

Transport(String)

An error occurred from the QUIC transport layer.

Websocket(String)

An error occurred from the websocket transport layer.

Networking(Error)

An error occurred from networking.

Io(String)

An error occurred from IO.

Configuration(String)

An error occurred with the provided configuration options.

Client(String)

An error occurred inside of the client.

CollectionNotFound

An attempt to use a Collection with a Database that it wasn’t defined within.

CollectionAlreadyDefined

A Collection being added already exists. This can be caused by a collection name not being unique.

DocumentNotFound(CollectionName, u64)

An attempt to update a document that doesn’t exist.

DocumentConflict(CollectionName, u64)

When updating a document, if a situation is detected where the contents have changed on the server since the Revision provided, a Conflict error will be returned.

UniqueKeyViolation

When saving a document in a collection with unique views, a document emits a key that is already emitted by an existing ocument, this error is returned.

Fields of UniqueKeyViolation

view: ViewName

The name of the view that the unique key violation occurred.

conflicting_document_id: u64

The document that caused the violation.

existing_document_id: u64

The document that already uses the same key.

InvalidName(InvalidNameError)

An invalid name was specified during schema creation.

PermissionDenied(PermissionDenied)

Permission was denied.

Password(String)

An internal error handling passwords was encountered.

UserNotFound

The user specified was not found. This will not be returned in response to an invalid username being used during login. It will be returned in other APIs that operate upon users.

InvalidCredentials

The credentials specified are not valid.